+++ app/src/main/java/org/oxycblt/auxio/detail/ArtistDetailFragment.kt
@@ -139,8 +139,8 @@ class ArtistDetailFragment :
     override fun onShowSortMenu(anchor: View) {
         menu(anchor, R.menu.menu_artist_sort) {
             val sort = detailModel.artistSort
-            requireNotNull(menu.findItem(sort.mode.itemId)).isChecked = true
-            requireNotNull(menu.findItem(R.id.option_sort_asc)).isChecked = sort.isAscending
+            unlikelyToBeNull(menu.findItem(sort.mode.itemId)).isChecked = true
+            unlikelyToBeNull(menu.findItem(R.id.option_sort_asc)).isChecked = sort.isAscending
             setOnMenuItemClickListener { item ->
                 item.isChecked = !item.isChecked
 
@@ -148,7 +148,7 @@ class ArtistDetailFragment :
                     if (item.itemId == R.id.option_sort_asc) {
                         sort.withAscending(item.isChecked)
                     } else {
-                        sort.withMode(requireNotNull(Sort.Mode.fromItemId(item.itemId)))
+                        sort.withMode(unlikelyToBeNull(Sort.Mode.fromItemId(item.itemId)))
                     }
 
                 true