+++ app/src/main/java/org/oxycblt/auxio/detail/GenreDetailFragment.kt
@@ -140,15 +140,15 @@ class GenreDetailFragment :
     override fun onShowSortMenu(anchor: View) {
         menu(anchor, R.menu.menu_genre_sort) {
             val sort = detailModel.genreSort
-            requireNotNull(menu.findItem(sort.mode.itemId)).isChecked = true
-            requireNotNull(menu.findItem(R.id.option_sort_asc)).isChecked = sort.isAscending
+            unlikelyToBeNull(menu.findItem(sort.mode.itemId)).isChecked = true
+            unlikelyToBeNull(menu.findItem(R.id.option_sort_asc)).isChecked = sort.isAscending
             setOnMenuItemClickListener { item ->
                 item.isChecked = !item.isChecked
                 detailModel.genreSort =
                     if (item.itemId == R.id.option_sort_asc) {
                         sort.withAscending(item.isChecked)
                     } else {
-                        sort.withMode(requireNotNull(Sort.Mode.fromItemId(item.itemId)))
+                        sort.withMode(unlikelyToBeNull(Sort.Mode.fromItemId(item.itemId)))
                     }
                 true
             }